requirements.txt
Django==4.2.6
djangorestframework==3.14.0
psycopg==3.1.12
psycopg2==2.9.9
